+++ Source/ACE.Entity/Enum/PortalWeenie.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace ACE.Entity.Enum
+{
+    [Flags]
+    public enum PortalWeenie : ushort
+    {
+        FloatingCity    = 8190,
+        GolemSanctum    = 7934,
+        HummingCrystal  = 9071,
+        Orphanage       = 27298
+    }
+}